+++ Doppel/Support/AppMetadata.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+public enum AppMetadata {
+    public static let name = "Doppel"
+    public static let version = "0.1.1"
+    public static let bundleIdentifier = "com.junowoz.doppel"
+    public static let releasesURL = URL(string: "https://github.com/junowoz/doppel/releases")!
+    public static let repositoryURL = URL(string: "https://github.com/junowoz/doppel")!
+}